+++ src/solver/pieces.rs
@@ -1,12 +1,12 @@
 pub trait ChessPiece {
-	fn relative_moves(&self) -> &[(i8, i8)];
+	fn relative_moves(&self) -> &[(i16, i16)];
 }
 
 #[derive(Copy, Clone)]
 pub struct StandardKnight;
 
 impl ChessPiece for StandardKnight {
-	fn relative_moves(&self) -> &[(i8, i8)] {
+	fn relative_moves(&self) -> &[(i16, i16)] {
 		&[
 			(2, 1),
 			(1, 2),
